+++ doc/version_history.rst
@@ -6,6 +6,43 @@
 Version History
 ###############
 
+v0.6.0
+======
+
+* Consolidated all multi-channel temperature topics into one.
+* Replaced the use of ts_salobj functions with ts_utils functions.
+* Added tests for all supported devices in the test class for the CSC.
+* Removed logging configuration from CSC run script.
+* Added telemetry for the computed dew point in all humidity sensors that don't provide it themselves.
+* Made sure that the CSC goes into FAULT state in case of an error.
+* Added location to the configuration of the devices.
+* Made sure that the CSC reports the sensor location in the telemetry.
+
+Requires:
+
+* ts_salobj 6.3
+* ts_idl 3.1
+* IDL file for ESS from ts_xml 10.1
+* ts_ess_common
+* ts_tcpip
+* ts_utils 1.0
+
+
+v0.5.1
+======
+
+* Fixed launch script to get index argument.
+* Added auto-enable capability.
+
+Requires:
+
+* ts_salobj 6.6
+* ts_idl 3.3
+* IDL file for ESS from ts_xml 10.0
+* ts_ess_controller
+* ts_ess_common
+* ts_tcpip
+
 v0.5.0
 ======